Split task details popup into new sections
