+++ frontend/src/functionality/Dashboard/stylesheets/TaskDetailsPopup/TaskDetailsPopup.css
